LPS-88869 Move dir
