+++ modules/apps/journal/journal-terms-of-use/src/main/java/com/liferay/journal/terms/of/use/internal/JournalArticleTermsOfUseContentProvider.java
@@ -14,7 +14,7 @@
 
 package com.liferay.journal.terms.of.use.internal;
 
-import com.liferay.portal.kernel.term.TermsOfUseContentProvider;
+import com.liferay.portal.kernel.terms.of.use.TermsOfUseContentProvider;
 
 import javax.servlet.RequestDispatcher;
 import javax.servlet.ServletContext;
