+++ modules/apps/portal-settings/portal-settings-web/src/main/resources/META-INF/resources/init.jsp
@@ -52,7 +52,7 @@ page import="com.liferay.portal.kernel.service.*" %><%@
 page import="com.liferay.portal.kernel.service.RoleLocalServiceUtil" %><%@
 page import="com.liferay.portal.kernel.servlet.taglib.DynamicInclude" %><%@
 page import="com.liferay.portal.kernel.servlet.taglib.ui.FormNavigatorConstants" %><%@
-page import="com.liferay.portal.kernel.term.TermsOfUseContentProvider" %><%@
+page import="com.liferay.portal.kernel.terms.of.use.TermsOfUseContentProvider" %><%@
 page import="com.liferay.portal.kernel.util.ArrayUtil" %><%@
 page import="com.liferay.portal.kernel.util.Constants" %><%@
 page import="com.liferay.portal.kernel.util.HtmlUtil" %><%@
