+++ portal-kernel/src/com/liferay/portal/kernel/terms/of/use/packageinfo
